Risolto problema con NumberUtils

+++ src/main/java/com/github/lory24/officinaio/utils/NumberUtils.java
@@ -7,7 +7,7 @@ public class NumberUtils {
 
     public boolean isNumber(String key) {
         try {
-            Integer.parseInt(key);
+            Double.parseDouble(key);
             return true;
         } catch (Exception ignored) {
             return false;
